refactor(channel): extract HTTP client helpers
